feat: split file field into a control and field composables
