+++ packages/core/src/useFileField/index.ts
@@ -1,3 +1,4 @@
+export * from './useFileControl';
 export * from './useFileField';
 export * from './useFileEntry';
 export * from './types';
